tracebacks of event handlers are now colorful
